pyproject.toml
[project]
name = "wordwield"
version = "0.1.0"
description = "Spell Execution Environment"
authors = [
    {name = "Alexander Podgorny", email = "alexanderpodgorny@gmail.com"}
]
dependencies = [
	"fastapi",
	"uvicorn",
	"pydantic",
	"httpx",
	"python-dotenv"
]

[project.scripts]
ww = "wordwield.__main__:main"

[tool.setuptools]
packages = ["wordwield"]
